LayoutTests/platform/win/plugins/npn-invalidate-rect-invalidates-window.html

<html>
<head>
<body>
<embed id="testPlugin" 
       type="application/x-webkit-test-netscape"
       src="data:text/plain,"
       test="npn-invalidate-rect-invalidates-window">
</embed>

<div>
    This tests that a plugin's HWND is invalidated by NPN_InvalidateRect.
</div>
<script>
  if (window.testRunner)
      testRunner.dumpAsText();
</script>
</body>
</html>
