LayoutTests/platform/chromium-win/plugins/npn-invalidate-rect-invalidates-window-expected.txt

CONSOLE MESSAGE: PLUGIN: Plugin's HWND has been invalidated as expected

This tests that a plugin's HWND is invalidated by NPN_InvalidateRect.
